feature: Added Blocked Recipe System to allow mass blocking of recipes of all types fix: Lumberaxes now have a config option to say which mods should prevent it from attempting whole tree destruction fix: Bombs, Mines, other entities, and systems now explode in all directions, instead of just towards positive coords. fix: small Material code changes

+++ src/main/java/minefantasy/mfr/material/CustomMaterial.java
@@ -156,7 +156,7 @@ public class CustomMaterial extends IForgeRegistryEntry.Impl<CustomMaterial>{
 		return craftTimeModifier;
 	}
 
-	public float getMeltingPoint() {
+	public Integer getMeltingPoint() {
 		return meltingPoint;
 	}
